Simplify filesystem data source
