+++ lib/nanoc/data_sources.rb
@@ -1,8 +1,8 @@
 # @api private
 module Nanoc::DataSources
-  autoload 'Filesystem',         'nanoc/data_sources/filesystem'
-  autoload 'FilesystemUnified',  'nanoc/data_sources/filesystem_unified'
-
-  Nanoc::DataSource.register '::Nanoc::DataSources::FilesystemUnified',  :filesystem_unified
-  Nanoc::DataSource.register '::Nanoc::DataSources::FilesystemUnified',  :filesystem
 end
+
+require_relative 'data_sources/filesystem'
+
+Nanoc::DataSource.register ::Nanoc::DataSources::Filesystem, :filesystem
+Nanoc::DataSource.register ::Nanoc::DataSources::Filesystem, :filesystem_unified
